requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
langchain-community
langchain-google-genai
PyPDF2